feat: Add OIDC authentication trough two different authentication flows
1. User (Interactively)
* Frontend requests auth server (CERN's keycloak) for public user token (insecure application), the user is always redirected to auth server login page and is required to login using his username, password and totp
* Frontend requests backend rest-api (django) for confidential user token by token exchange (secure application), the backend always check the public user token validity and if was sent from public application
* Any request sent to backend rest-api puts the confidential user access token and the backend always check the token validity, audience and authorized parties
* Right before confidential user token expires (stored in frontend localStorage) the frontend renew the public user token and requests a new exchange

2. Machine-to-Machine (Non-interactively)
* The user can request an api access token from backend through `issue-api-token` route, sending an allowed application client secret key (think of it as an api key)
* For subsequent requests the user must send the received api access token and always requests a new one when the older one expires

Note: Authentication flow #2 is machete (gambiarra, kludge, τσαπατσουλιά /tsapatsoulia/) because the auth server should link an api key for each user instead of "linking" multiple users to an application through e-groups. That is, for fine-grained api access roles instead of giving a role to each user we need to create an application to each user (this is inefficient because it will generate multiple applications) or group multiple users in an application (this is insecure because multiple users would share the same client secret key).

Note: Squashed commit from branch `feat/add-cern-sso-authentication` (originally 19 commits)