fix: add `file_uuid` to `FileIndex` model and switch the unique constraint from `file_path` to `file_uuid`
Using the file path for uniquely identifying an file is not a good idea, because if you change the storage location all the files will be indexed again. A good idea would be to compute the file hash but that consumes a lot of time for larger files, hopefully each ROOT file has an associated UUID and it can be extract really fast by reading the TFile Header.

Since the application is still in development stage and all data can be processed again, it was easier to delete the old migrations and generate a new one. Otherwise we would need to define a default to the new field `file_uuid`.