fix: authentication div showing when user authenticated in a window but opened the page again in another window
When the user opens the page in another window but it was previously authenticated from another tab the `tokenExchange` will not be set to true by the event capture, because the event was captured in another window. So, whenever the root component loads we need to check if the confidential token is present in the localStorage to set the `tokenExchange` variable to true.