refactor: remove exchange_token endpoint
This commit removes the `exchange-token` endpoint from the `cern_auth`
router, that was used only by the frontend to fetch a token for the
backend application client of the dials api.

Since this endpoint was also resolving the user default workspace, a new
GET endpoint was added `use-default-workspace`. Therefore the frontend
can still automatically configure the user pre-selected workpspace based
on his e-groups configured on the backend application client.