feat: secure the entire application using keycloak-js
This commit refactors all components and API services to fetch
microservices credentials from the newly added `KeycloakExchange`
component, requesting a token exchange directly to the Keycloak API.

The login procedure is completely refactored using only `keycloak-js`
and the custom `KeycloakWrapper` component. For the users, it will not
require a button click in the `Sign in with CERN` since the
`KeycloakWrapper` will start the login procedure automatically as soon
as the page loads.

Furthermore, the `PrivateRoute` component is deprecated since the entire
component is now secured behind the SSO. Likewise the `AppNavbar`
component doesn't need to check if the user is authenticated anymore.